move stores to the app folder
